+++ ostap/io/bz2shelve.py
@@ -143,12 +143,6 @@ __all__ = (
     'tmpdb'    ,   ## create TEMPORARY data base 
     )
 # =============================================================================
-from ostap.logger.logger import getLogger
-if '__main__' == __name__ : logger = getLogger ( 'ostap.io.bz2shelve' )
-else                      : logger = getLogger ( __name__             )
-# =============================================================================
-logger.debug ( "Simple generic (c)Pickle-based ``bzip2''-database"    )
-# =============================================================================
 from sys import version_info as python_version 
 # =============================================================================
 try:
@@ -172,7 +166,14 @@ import os, sys
 import bz2         ## use bz2 to compress DB-content 
 import shelve      ## 
 import shutil
-from ostap.io.compress_shelve import CompressShelf
+from   ostap.io.compress_shelve import CompressShelf
+from   ostap.io.dbase           import TmpDB 
+# =============================================================================
+from ostap.logger.logger import getLogger
+if '__main__' == __name__ : logger = getLogger ( 'ostap.io.bz2shelve' )
+else                      : logger = getLogger ( __name__             )
+# =============================================================================
+logger.debug ( "Simple generic (c)Pickle-based ``bzip2''-database"    )
 # =============================================================================
 ## @class Bz2Shelf
 #  ``Bzip2''-version of ``shelve''-database
@@ -364,41 +365,37 @@ def open ( filename                  ,
 #  TEMPORARY bzip2-version of ``shelve''-database
 #  @author Vanya BELYAEV Ivan.Belyaev@cern.ch
 #  @date   2015-10-31
-class TmpBz2Shelf(Bz2Shelf):
+class TmpBz2Shelf(Bz2Shelf,TmpDB):
     """
     TEMPORARY ``bzip2''-version of ``shelve''-database     
     """    
-    def __init__(
-        self                           ,
-        protocol    = HIGHEST_PROTOCOL , 
-        compress    = 9                ,
-        silent      = False            ,
-        keyencoding = ENCODING         ) :
+    def __init__( self                           ,
+                  protocol    = HIGHEST_PROTOCOL , 
+                  compress    = 9                ,
+                  silent      = False            ,
+                  keyencoding = ENCODING         ,
+                  remove      = True             ,
+                  keep        = False            ) :
 
-        ## create temporary file name 
-        import ostap.utils.cleanup as CU 
-        filename = CU.CleanUp.tempfile ( prefix = 'ostap-tmpdb-' , suffix = '.bz2db' )
+        ## initialize the base: generate the name 
+        TmpDB.__init__ ( self , suffix = '.bz2db' , remove = remove , keep = keep ) 
         
-        Bz2Shelf.__init__ ( self        ,  
-                            filename    ,
-                            'c'         ,
-                            protocol    ,
-                            compress    , 
-                            False       , ## writeback 
-                            silent      ,
-                            keyencoding ) 
+        ## open DB 
+        Bz2Shelf.__init__ ( self          ,  
+                            self.tmp_name ,
+                            'c'           ,
+                            protocol      ,
+                            compress      , 
+                            False         , ## writeback 
+                            silent        ,
+                            keyencoding   ) 
         
     ## close and delete the file 
     def close ( self )  :
         ## close the shelve file
         Bz2Shelf.close ( self )
-        ## delete the file 
-        fname = self.nominal_dbname 
-        if os.path.exists ( fname ) :
-            try :
-                os.unlink ( fname )
-            except : 
-                pass
+        ## delete the file
+        TmpDB.clean    ( self )
             
 # =============================================================================
 ## helper function to open TEMPORARY ZipShelve data base#
@@ -407,7 +404,9 @@ class TmpBz2Shelf(Bz2Shelf):
 def tmpdb ( protocol      = HIGHEST_PROTOCOL ,
             compresslevel = 9                , 
             silent        = True             ,
-            keyencoding   = ENCODING         ) :
+            keyencoding   = ENCODING         ,
+            remove        = True             ,   ## immediate remove 
+            keep          = False            ) : ## keep it 
     """Open a TEMPORARY persistent dictionary for reading and writing.
     
     The optional protocol parameter specifies the
@@ -418,8 +417,9 @@ def tmpdb ( protocol      = HIGHEST_PROTOCOL ,
     return TmpBz2Shelf ( protocol      ,
                          compresslevel ,
                          silent        ,
-                         keyencoding   ) 
-    
+                         keyencoding   ,
+                         remove        ,
+                         keep          ) 
     
 # =============================================================================
 if '__main__' == __name__ :